# eslint perf improvement