--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -40,6 +40,7 @@ module.exports = {
     '@metamask/eslint-config/config/mocha',
     'plugin:react/recommended',
     'plugin:react-hooks/recommended',
+    'prettier',
   ],
 
   plugins: ['@babel', 'react', 'import', 'prettier'],
@@ -53,59 +54,6 @@ module.exports = {
     // Prettier changes and reasoning
 
     'prettier/prettier': 'error',
-
-    // Our eslint config has the default setting for this as error. This
-    // include beforeBlockComment: true, but in order to match the prettier
-    // spec you have to enable before and after blocks, objects and arrays
-    // https://github.com/prettier/eslint-config-prettier#lines-around-comment
-    'lines-around-comment': [
-      'error',
-      {
-        beforeBlockComment: true,
-        afterLineComment: false,
-        allowBlockStart: true,
-        allowBlockEnd: true,
-        allowObjectStart: true,
-        allowObjectEnd: true,
-        allowArrayStart: true,
-        allowArrayEnd: true,
-      },
-    ],
-    // Prettier has some opinions on mixed-operators, and there is ongoing work
-    // to make the output code clear. It is better today then it was when the first
-    // PR to add prettier. That being said, the workaround for keeping this rule enabled
-    // requires breaking parts of operations into different variables -- which I believe
-    // to be worse. https://github.com/prettier/eslint-config-prettier#no-mixed-operators
-    'no-mixed-operators': 'off',
-    // Prettier wraps single line functions with ternaries, etc in parens by default, but
-    // if the line is long enough it breaks it into a separate line and removes the parens.
-    // The second behavior conflicts with this rule. There is some guides on the repo about
-    // how you can keep it enabled:
-    // https://github.com/prettier/eslint-config-prettier#no-confusing-arrow
-    // However, in practice this conflicts with prettier adding parens around short lines,
-    // when autofixing in vscode and others.
-    'no-confusing-arrow': 'off',
-    // There is no configuration in prettier for how it stylizes regexes, which conflicts
-    // with wrap-regex.
-    'wrap-regex': 'off',
-    // Prettier handles all indentation automagically. it can be configured here
-    // https://prettier.io/docs/en/options.html#tab-width but the default matches our
-    // style.
-    indent: 'off',
-    // This rule conflicts with the way that prettier breaks code across multiple lines when
-    // it exceeds the maximum length. Prettier optimizes for readability while simultaneously
-    // maximizing the amount of code per line.
-    'function-paren-newline': 'off',
-    // This rule throws an error when there is a line break in an arrow function declaration
-    // but prettier breaks arrow function declarations to be as readable as possible while
-    // still conforming to the width rules.
-    'implicit-arrow-linebreak': 'off',
-    // This rule would result in an increase in white space in lines with generator functions,
-    // which impacts prettier's goal of maximizing code per line and readability. There is no
-    // current workaround.
-    'generator-star-spacing': 'off',
-    'default-param-last': 'off',
-    'require-atomic-updates': 'off',
     'import/no-unassigned-import': 'off',
     'prefer-object-spread': 'error',
     'react/no-unused-prop-types': 'error',
@@ -115,28 +63,11 @@ module.exports = {
       'error',
       { props: 'never', children: 'never' },
     ],
-    'react/jsx-equals-spacing': 'error',
     'react/no-deprecated': 'error',
     'react/default-props-match-prop-types': 'error',
-    'react/jsx-closing-tag-location': [
-      'error',
-      { selfClosing: 'tag-aligned', nonEmpty: 'tag-aligned' },
-    ],
     'react/jsx-no-duplicate-props': 'error',
-    'react/jsx-closing-bracket-location': 'error',
-    'react/jsx-first-prop-new-line': ['error', 'multiline'],
-    'react/jsx-max-props-per-line': [
-      'error',
-      { maximum: 1, when: 'multiline' },
-    ],
-    'react/jsx-tag-spacing': [
-      'error',
-      {
-        closingSlash: 'never',
-        beforeSelfClosing: 'always',
-        afterOpening: 'never',
-      },
-    ],
+    'default-param-last': 'off',
+    'require-atomic-updates': 'off',
 
     'no-invalid-this': 'off',
     '@babel/no-invalid-this': 'error',
@@ -177,6 +108,26 @@ module.exports = {
       rules: {
         // Mocha will re-assign `this` in a test context
         '@babel/no-invalid-this': 'off',
+      },
+    },
+    {
+      files: ['**/!(*.test).js'],
+      rules: {
+        'mocha/max-top-level-suites': 'off',
+        'mocha/no-identical-tests': 'off',
+        'mocha/no-nested-tests': 'off',
+        'mocha/no-identical-title': 'off',
+        'mocha/no-skipped-tests': 'off',
+        'mocha/no-exclusive-tests': 'off',
+        'mocha/no-hooks-for-single-case': 'off',
+        'mocha/no-async-describe': 'off',
+        'mocha/no-sibling-hooks': 'off',
+        'mocha/no-global-tests': 'off',
+        'mocha/no-pending-tests': 'off',
+        'mocha/no-mocha-arrows': 'off',
+        'mocha/no-top-level-hooks': 'off',
+        'mocha/handle-done-callback': 'off',
+        'mocha/no-return-and-callback': 'off',
       },
     },
     {