Develop: 154.76s total. 31.1 seconds for prettier, 11.42 seconds for import/namespace, 48.82 seconds for the top mocha rules? 
```
➜ TIMING=1 yarn eslint . --ext js
yarn run v1.22.4
$ /Users/braddecker/Dev/consensys/metamask-extension/node_modules/.bin/eslint . --ext js
Rule                           | Time (ms) | Relative
:------------------------------|----------:|--------:
prettier/prettier              | 31408.751 |    24.1%
import/namespace               | 11417.413 |     8.8%
mocha/no-nested-tests          | 10285.072 |     7.9%
mocha/no-identical-title       |  8847.269 |     6.8%
mocha/no-skipped-tests         |  6528.291 |     5.0%
mocha/no-exclusive-tests       |  6485.040 |     5.0%
mocha/no-hooks-for-single-case |  6025.373 |     4.6%
mocha/max-top-level-suites     |  3739.456 |     2.9%
mocha/no-async-describe        |  3454.342 |     2.7%
mocha/no-global-tests          |  3441.664 |     2.6%
✨  Done in 154.76s.
```

This PR: 
Develop: 93.55s total. 11.36 seconds for prettier, 11.39 seconds for import/namespace, 17.18 seconds for the top mocha rules

```
➜ TIMING=1 yarn eslint . --ext js
yarn run v1.22.4
$ /Users/braddecker/Dev/consensys/metamask-extension/node_modules/.bin/eslint . --ext js
Rule                           | Time (ms) | Relative
:------------------------------|----------:|--------:
import/namespace               | 11392.980 |    16.2%
prettier/prettier              | 11359.871 |    16.1%
mocha/no-nested-tests          |  4515.200 |     6.4%
mocha/no-identical-title       |  4025.436 |     5.7%
mocha/no-hooks-for-single-case |  2907.527 |     4.1%
mocha/no-exclusive-tests       |  2868.671 |     4.1%
mocha/no-skipped-tests         |  2854.651 |     4.0%
node/no-extraneous-import      |  2708.008 |     3.8%
node/no-restricted-import      |  2447.184 |     3.5%
import/named                   |  1739.707 |     2.5%
✨  Done in 93.55s.
```